# Sorting merge data corruption