This adds a unit test that recreates the failure seen in #139 

The merged row reader will overwrite values in its internal buffers when filling up the buffer during the `ReadRows` call. 
This causes data corruption in the final `rows` that are returned to the caller.

This adds a potential fix for this issue by called the `rowAllocator` capture method we can copy in the byte arrays. However this creates some significant hits to the benchmarks seen below. Is there a better solution to prevent this buffer corruption? 

```bash
name                                                        old time/op    new time/op    delta
MergeRowGroups/BOOLEAN/groups=1,rows=10000-10                 20.7µs ± 0%    23.8µs ± 0%    +14.79%  (p=0.000 n=9+9)
MergeRowGroups/BOOLEAN/groups=2,rows=20000-10                 28.4µs ± 0%    31.9µs ± 0%    +12.32%  (p=0.000 n=10+10)
MergeRowGroups/BOOLEAN/groups=3,rows=30000-10                 36.4µs ± 0%    40.0µs ± 0%     +9.85%  (p=0.000 n=9+8)
MergeRowGroups/INT32/groups=1,rows=10000-10                   17.8µs ± 0%    21.1µs ± 1%    +18.59%  (p=0.000 n=8+9)
MergeRowGroups/INT32/groups=2,rows=20000-10                   32.9µs ± 0%    35.7µs ± 0%     +8.31%  (p=0.000 n=9+9)
MergeRowGroups/INT32/groups=3,rows=30000-10                   43.2µs ± 0%    46.3µs ± 0%     +6.98%  (p=0.000 n=10+9)
MergeRowGroups/INT64/groups=1,rows=10000-10                   17.8µs ± 0%    21.0µs ± 0%    +17.86%  (p=0.000 n=9+10)
MergeRowGroups/INT64/groups=2,rows=20000-10                   32.9µs ± 0%    35.6µs ± 0%     +8.18%  (p=0.000 n=9+10)
MergeRowGroups/INT64/groups=3,rows=30000-10                   43.3µs ± 0%    46.3µs ± 0%     +6.91%  (p=0.000 n=10+10)
MergeRowGroups/INT96/groups=1,rows=10000-10                   29.6µs ± 0%    32.8µs ± 0%    +10.76%  (p=0.000 n=10+10)
MergeRowGroups/INT96/groups=2,rows=20000-10                   56.4µs ± 0%    59.3µs ± 0%     +5.10%  (p=0.000 n=10+7)
MergeRowGroups/INT96/groups=3,rows=30000-10                   78.2µs ± 0%    81.2µs ± 0%     +3.86%  (p=0.000 n=9+10)
MergeRowGroups/FLOAT/groups=1,rows=10000-10                   17.9µs ± 0%    21.1µs ± 0%    +17.76%  (p=0.000 n=10+8)
MergeRowGroups/FLOAT/groups=2,rows=20000-10                   36.7µs ± 0%    39.6µs ± 0%     +7.89%  (p=0.000 n=10+10)
MergeRowGroups/FLOAT/groups=3,rows=30000-10                   50.4µs ± 0%    53.5µs ± 0%     +6.09%  (p=0.000 n=9+10)
MergeRowGroups/DOUBLE/groups=1,rows=10000-10                  18.0µs ± 0%    21.1µs ± 0%    +17.48%  (p=0.000 n=10+10)
MergeRowGroups/DOUBLE/groups=2,rows=20000-10                  36.7µs ± 0%    39.6µs ± 0%     +7.90%  (p=0.000 n=9+8)
MergeRowGroups/DOUBLE/groups=3,rows=30000-10                  50.5µs ± 0%    53.5µs ± 0%     +6.00%  (p=0.000 n=10+9)
MergeRowGroups/BYTE_ARRAY/groups=1,rows=10000-10              18.7µs ± 0%    36.7µs ± 0%    +96.32%  (p=0.000 n=9+9)
MergeRowGroups/BYTE_ARRAY/groups=2,rows=20000-10              42.9µs ± 0%    56.9µs ± 2%    +32.56%  (p=0.000 n=10+9)
MergeRowGroups/BYTE_ARRAY/groups=3,rows=30000-10              60.9µs ± 0%    76.9µs ± 0%    +26.16%  (p=0.000 n=9+9)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=1,rows=10000-10    18.3µs ± 0%    27.6µs ± 1%    +51.11%  (p=0.000 n=10+9)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=2,rows=20000-10    39.4µs ± 1%    48.8µs ± 0%    +23.81%  (p=0.000 n=10+10)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=3,rows=30000-10    56.9µs ± 1%    66.4µs ± 0%    +16.61%  (p=0.000 n=10+9)
MergeRowGroups/STRING/groups=1,rows=10000-10                  18.8µs ± 1%    27.4µs ± 0%    +45.35%  (p=0.000 n=9+9)
MergeRowGroups/STRING/groups=2,rows=20000-10                  30.0µs ± 0%    38.4µs ± 0%    +27.79%  (p=0.000 n=10+10)
MergeRowGroups/STRING/groups=3,rows=30000-10                  41.7µs ± 0%    50.3µs ± 0%    +20.64%  (p=0.000 n=10+10)
MergeRowGroups/STRING_(dict)/groups=1,rows=10000-10           18.6µs ± 1%    27.2µs ± 0%    +46.57%  (p=0.000 n=10+10)
MergeRowGroups/STRING_(dict)/groups=2,rows=20000-10           29.8µs ± 0%    38.2µs ± 0%    +28.52%  (p=0.000 n=9+10)
MergeRowGroups/STRING_(dict)/groups=3,rows=30000-10           41.4µs ± 0%    50.3µs ± 0%    +21.57%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=1,rows=10000-10                    18.3µs ± 0%    29.2µs ± 0%    +59.60%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=2,rows=20000-10                    36.8µs ± 0%    47.6µs ± 0%    +29.53%  (p=0.000 n=10+9)
MergeRowGroups/UUID/groups=3,rows=30000-10                    50.1µs ± 0%    60.3µs ± 0%    +20.42%  (p=0.000 n=9+10)
MergeRowGroups/time.Time/groups=1,rows=10000-10               17.9µs ± 0%    21.0µs ± 0%    +17.39%  (p=0.000 n=10+10)
MergeRowGroups/time.Time/groups=2,rows=20000-10               35.1µs ± 0%    37.9µs ± 0%     +7.92%  (p=0.000 n=10+9)
MergeRowGroups/time.Time/groups=3,rows=30000-10               47.6µs ± 0%    50.5µs ± 0%     +6.17%  (p=0.000 n=9+10)
MergeRowGroups/time.Time_in_ms/groups=1,rows=10000-10         17.9µs ± 0%    21.0µs ± 0%    +17.62%  (p=0.000 n=10+10)
MergeRowGroups/time.Time_in_ms/groups=2,rows=20000-10         35.0µs ± 0%    37.8µs ± 0%     +8.08%  (p=0.000 n=10+10)
MergeRowGroups/time.Time_in_ms/groups=3,rows=30000-10         47.6µs ± 0%    50.5µs ± 0%     +6.09%  (p=0.000 n=9+9)
MergeRowGroups/DECIMAL/groups=1,rows=10000-10                 17.9µs ± 0%    21.0µs ± 0%    +17.31%  (p=0.000 n=10+9)
MergeRowGroups/DECIMAL/groups=2,rows=20000-10                 37.8µs ± 0%    40.3µs ± 0%     +6.84%  (p=0.000 n=8+8)

name                                                        old row/s      new row/s      delta
MergeRowGroups/BOOLEAN/groups=1,rows=10000-10                  48.2M ± 0%     42.0M ± 0%    -12.88%  (p=0.000 n=9+9)
MergeRowGroups/BOOLEAN/groups=2,rows=20000-10                  35.2M ± 0%     31.4M ± 0%    -10.97%  (p=0.000 n=10+10)
MergeRowGroups/BOOLEAN/groups=3,rows=30000-10                  27.5M ± 0%     25.0M ± 0%     -8.96%  (p=0.000 n=9+8)
MergeRowGroups/INT32/groups=1,rows=10000-10                    56.2M ± 0%     47.4M ± 1%    -15.68%  (p=0.000 n=8+9)
MergeRowGroups/INT32/groups=2,rows=20000-10                    30.4M ± 0%     28.0M ± 0%     -7.67%  (p=0.000 n=9+9)
MergeRowGroups/INT32/groups=3,rows=30000-10                    23.1M ± 0%     21.6M ± 0%     -6.53%  (p=0.000 n=10+9)
MergeRowGroups/INT64/groups=1,rows=10000-10                    56.1M ± 0%     47.6M ± 0%    -15.16%  (p=0.000 n=9+10)
MergeRowGroups/INT64/groups=2,rows=20000-10                    30.4M ± 0%     28.1M ± 0%     -7.56%  (p=0.000 n=9+10)
MergeRowGroups/INT64/groups=3,rows=30000-10                    23.1M ± 0%     21.6M ± 0%     -6.46%  (p=0.000 n=10+10)
MergeRowGroups/INT96/groups=1,rows=10000-10                    33.8M ± 0%     30.5M ± 0%     -9.72%  (p=0.000 n=10+10)
MergeRowGroups/INT96/groups=2,rows=20000-10                    17.7M ± 0%     16.9M ± 0%     -4.85%  (p=0.000 n=10+7)
MergeRowGroups/INT96/groups=3,rows=30000-10                    12.8M ± 0%     12.3M ± 0%     -3.72%  (p=0.000 n=9+10)
MergeRowGroups/FLOAT/groups=1,rows=10000-10                    55.8M ± 0%     47.4M ± 0%    -15.11%  (p=0.000 n=10+9)
MergeRowGroups/FLOAT/groups=2,rows=20000-10                    27.2M ± 0%     25.3M ± 0%     -7.31%  (p=0.000 n=10+10)
MergeRowGroups/FLOAT/groups=3,rows=30000-10                    19.8M ± 0%     18.7M ± 0%     -5.74%  (p=0.000 n=9+10)
MergeRowGroups/DOUBLE/groups=1,rows=10000-10                   55.6M ± 0%     47.3M ± 0%    -14.88%  (p=0.000 n=10+10)
MergeRowGroups/DOUBLE/groups=2,rows=20000-10                   27.2M ± 0%     25.2M ± 0%     -7.32%  (p=0.000 n=9+8)
MergeRowGroups/DOUBLE/groups=3,rows=30000-10                   19.8M ± 0%     18.7M ± 0%     -5.66%  (p=0.000 n=10+9)
MergeRowGroups/BYTE_ARRAY/groups=1,rows=10000-10               53.5M ± 0%     27.2M ± 0%    -49.06%  (p=0.000 n=9+9)
MergeRowGroups/BYTE_ARRAY/groups=2,rows=20000-10               23.3M ± 0%     17.6M ± 2%    -24.56%  (p=0.000 n=10+9)
MergeRowGroups/BYTE_ARRAY/groups=3,rows=30000-10               16.4M ± 0%     13.0M ± 0%    -20.74%  (p=0.000 n=9+9)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=1,rows=10000-10     54.8M ± 0%     36.3M ± 1%    -33.82%  (p=0.000 n=10+9)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=2,rows=20000-10     25.4M ± 1%     20.5M ± 0%    -19.23%  (p=0.000 n=10+10)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=3,rows=30000-10     17.6M ± 1%     15.1M ± 0%    -14.24%  (p=0.000 n=10+9)
MergeRowGroups/STRING/groups=1,rows=10000-10                   53.1M ± 1%     36.5M ± 0%    -31.20%  (p=0.000 n=9+9)
MergeRowGroups/STRING/groups=2,rows=20000-10                   33.3M ± 0%     26.1M ± 0%    -21.74%  (p=0.000 n=10+10)
MergeRowGroups/STRING/groups=3,rows=30000-10                   24.0M ± 0%     19.9M ± 0%    -17.11%  (p=0.000 n=10+10)
MergeRowGroups/STRING_(dict)/groups=1,rows=10000-10            53.8M ± 1%     36.7M ± 0%    -31.77%  (p=0.000 n=10+10)
MergeRowGroups/STRING_(dict)/groups=2,rows=20000-10            33.6M ± 0%     26.2M ± 0%    -22.19%  (p=0.000 n=9+10)
MergeRowGroups/STRING_(dict)/groups=3,rows=30000-10            24.2M ± 0%     19.9M ± 0%    -17.74%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=1,rows=10000-10                     54.7M ± 0%     34.3M ± 0%    -37.35%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=2,rows=20000-10                     27.2M ± 0%     21.0M ± 0%    -22.80%  (p=0.000 n=10+9)
MergeRowGroups/UUID/groups=3,rows=30000-10                     20.0M ± 0%     16.6M ± 0%    -16.96%  (p=0.000 n=9+10)
MergeRowGroups/time.Time/groups=1,rows=10000-10                55.9M ± 0%     47.6M ± 0%    -14.81%  (p=0.000 n=10+10)
MergeRowGroups/time.Time/groups=2,rows=20000-10                28.5M ± 0%     26.4M ± 0%     -7.34%  (p=0.000 n=10+9)
MergeRowGroups/time.Time/groups=3,rows=30000-10                21.0M ± 0%     19.8M ± 0%     -5.82%  (p=0.000 n=9+10)
MergeRowGroups/time.Time_in_ms/groups=1,rows=10000-10          56.0M ± 0%     47.6M ± 0%    -14.98%  (p=0.000 n=10+10)
MergeRowGroups/time.Time_in_ms/groups=2,rows=20000-10          28.6M ± 0%     26.4M ± 0%     -7.48%  (p=0.000 n=10+10)
MergeRowGroups/time.Time_in_ms/groups=3,rows=30000-10          21.0M ± 0%     19.8M ± 0%     -5.74%  (p=0.000 n=9+9)
MergeRowGroups/DECIMAL/groups=1,rows=10000-10                  55.9M ± 0%     47.7M ± 0%    -14.76%  (p=0.000 n=10+9)
MergeRowGroups/DECIMAL/groups=2,rows=20000-10                  26.5M ± 0%     24.8M ± 0%     -6.40%  (p=0.000 n=8+8)

name                                                        old alloc/op   new alloc/op   delta
MergeRowGroups/BOOLEAN/groups=1,rows=10000-10                   498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/BOOLEAN/groups=2,rows=20000-10                   493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/BOOLEAN/groups=3,rows=30000-10                   493B ± 0%      492B ± 0%       ~     (p=0.650 n=10+10)
MergeRowGroups/INT32/groups=1,rows=10000-10                     498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/INT32/groups=2,rows=20000-10                     493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/INT32/groups=3,rows=30000-10                     493B ± 0%      493B ± 0%       ~     (p=0.294 n=10+8)
MergeRowGroups/INT64/groups=1,rows=10000-10                     498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/INT64/groups=2,rows=20000-10                     493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/INT64/groups=3,rows=30000-10                     492B ± 0%      493B ± 0%     +0.12%  (p=0.015 n=10+9)
MergeRowGroups/INT96/groups=1,rows=10000-10                   16.5kB ± 0%    16.5kB ± 0%       ~     (all equal)
MergeRowGroups/INT96/groups=2,rows=20000-10                   16.5kB ± 0%    16.5kB ± 0%       ~     (all equal)
MergeRowGroups/INT96/groups=3,rows=30000-10                   16.5kB ± 0%    16.5kB ± 0%       ~     (all equal)
MergeRowGroups/FLOAT/groups=1,rows=10000-10                     498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/FLOAT/groups=2,rows=20000-10                     493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/FLOAT/groups=3,rows=30000-10                     493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/DOUBLE/groups=1,rows=10000-10                    498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/DOUBLE/groups=2,rows=20000-10                    493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/DOUBLE/groups=3,rows=30000-10                    493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/BYTE_ARRAY/groups=1,rows=10000-10                498B ± 0%    26305B ± 0%  +5182.13%  (p=0.000 n=10+10)
MergeRowGroups/BYTE_ARRAY/groups=2,rows=20000-10                493B ± 0%    26509B ± 0%  +5277.06%  (p=0.000 n=10+10)
MergeRowGroups/BYTE_ARRAY/groups=3,rows=30000-10                493B ± 0%    35308B ± 0%  +7061.87%  (p=0.000 n=10+8)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=1,rows=10000-10      498B ± 0%    13196B ± 0%  +2549.88%  (p=0.000 n=10+10)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=2,rows=20000-10      493B ± 0%    13396B ± 0%  +2617.30%  (p=0.000 n=10+10)
MergeRowGroups/FIXED_LEN_BYTE_ARRAY/groups=3,rows=30000-10      493B ± 0%    17834B ± 0%  +3517.38%  (p=0.000 n=10+10)
MergeRowGroups/STRING/groups=1,rows=10000-10                    498B ± 0%     6642B ± 0%  +1233.82%  (p=0.000 n=10+10)
MergeRowGroups/STRING/groups=2,rows=20000-10                    493B ± 0%     6841B ± 0%  +1287.71%  (p=0.000 n=10+10)
MergeRowGroups/STRING/groups=3,rows=30000-10                    493B ± 0%     9091B ± 0%  +1745.55%  (p=0.000 n=10+10)
MergeRowGroups/STRING_(dict)/groups=1,rows=10000-10             498B ± 0%     6642B ± 0%  +1233.73%  (p=0.000 n=10+8)
MergeRowGroups/STRING_(dict)/groups=2,rows=20000-10             493B ± 0%     6841B ± 0%  +1287.69%  (p=0.000 n=10+10)
MergeRowGroups/STRING_(dict)/groups=3,rows=30000-10             493B ± 0%     9092B ± 0%  +1744.24%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=1,rows=10000-10                      498B ± 0%    26304B ± 0%  +5181.87%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=2,rows=20000-10                      493B ± 0%    26503B ± 0%  +5275.78%  (p=0.000 n=10+10)
MergeRowGroups/UUID/groups=3,rows=30000-10                      493B ± 0%    17836B ± 0%  +3517.85%  (p=0.000 n=10+10)
MergeRowGroups/time.Time/groups=1,rows=10000-10                 498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/time.Time/groups=2,rows=20000-10                 493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/time.Time/groups=3,rows=30000-10                 493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/time.Time_in_ms/groups=1,rows=10000-10           498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/time.Time_in_ms/groups=2,rows=20000-10           493B ± 0%      493B ± 0%       ~     (all equal)
MergeRowGroups/time.Time_in_ms/groups=3,rows=30000-10           493B ± 0%      493B ± 0%       ~     (p=0.248 n=10+9)
MergeRowGroups/DECIMAL/groups=1,rows=10000-10                   498B ± 0%      498B ± 0%       ~     (all equal)
MergeRowGroups/DECIMAL/groups=2,rows=20000-10                   493B ± 0%      493B ± 0%       ~     (all equal)
```