@thorfour 2024-05-29
> Hello @thorfour, thanks for taking the initiative to put together a fix.
> 
> I wonder if we could solve this problem without introducing more copies; since we know that each `ReadRows` call on the underlying `RowReader` will read a batch of values that remain valid until the next `ReadRows` call, we could break out of the `mergedRowReader.ReadRows` method each time we exhaust one of the sub-buffers that reads from each underlying reader. This would ensure that we are not invalidating pointers internally before returning from `ReadRows`.

Yea let me see how well that works