@thorfour 2024-05-30
@achille-roussel updated by having the merged row reader return early if the buffer is exhausted. It's definitely better performing than previously.

[diff.txt](https://github.com/parquet-go/parquet-go/files/15504476/diff.txt)