@achille-roussel 2024-05-30
Thanks for making the change, the impact looks within the ballpark of what I would expect; interesting to see that it reduces memory utilization in those benchmarks as well.

Now trying to understand what's going on with CI... looks like a dependency broke our build?

```
Error:  The build could not read 1 project -> [Help 1]
Error:    
Error:    The project  (/home/runner/work/parquet-go/parquet-go/parquet-mr-apache-parquet-1.13.1/parquet-cli/pom.xml) has 1 error
Error:      Non-readable POM /home/runner/work/parquet-go/parquet-go/parquet-mr-apache-parquet-1.13.1/parquet-cli/pom.xml: /home/runner/work/parquet-go/parquet-go/parquet-mr-apache-parquet-1.13.1/parquet-cli/pom.xml (No such file or directory)
Error:  
Error:  To see the full stack trace of the errors, re-run Maven with the -e switch.
Error:  Re-run Maven using the -X switch to enable full debug logging.
Error:  
Error:  For more information about the errors and possible solutions, please read the following articles:
Error:  [Help 1] http://cwiki.apache.org/confluence/display/MAVEN/ProjectBuildingException
Error: Process completed with exit code 1.
```