@thorfour 2024-05-30
> Now trying to understand what's going on with CI... looks like a dependency broke our build?

Yea I had no idea what was going on with CI. All the tests passed locally.