@thorfour 2024-05-30
> See https://github.com/parquet-go/parquet-go/pull/141 for a fix @thorfour 

Thanks, all green now. Let me know if this is good to merge