@MichaHoffmann 2024-11-26
I still got corrupted rows from a MergedRowReader in 0.24.0 which should have this patch. I fixed internally by calling "Clone()" on the "head()" as @fpetkovski described here https://github.com/segmentio/parquet-go/pull/512; but I think this bug is still happening on 0.24.0